move sFnControllerConfiguration to new file
