+++ internal/state/state.go
@@ -1,6 +1,8 @@
 package state
 
 import (
+	"github.com/kyma-project/serverless-manager/api/v1alpha1"
+	"k8s.io/client-go/tools/record"
 	"time"
 
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -34,4 +36,28 @@ func requeueAfter(duration time.Duration) (stateFn, *ctrl.Result, error) {
 	return nil, &ctrl.Result{
 		RequeueAfter: duration,
 	}, nil
+}
+
+type fieldsToUpdate []struct {
+	specField   string
+	statusField *string
+	fieldName   string
+}
+
+func updateStatusFields(eventRecorder record.EventRecorder, instance *v1alpha1.Serverless, fields fieldsToUpdate) {
+	for _, field := range fields {
+		if field.specField != *field.statusField {
+			oldStatusValue := *field.statusField
+			*field.statusField = field.specField
+			eventRecorder.Eventf(
+				instance,
+				"Normal",
+				string(v1alpha1.ConditionReasonConfiguration),
+				"%s set from '%s' to '%s'",
+				field.fieldName,
+				oldStatusValue,
+				field.specField,
+			)
+		}
+	}
 }
